Do not hide figure environments from translation
In the original version, every \begin{figure} ... \end{figure} environment was replaced by a token as a whole, thus being hidden from language translation. Typically, this is not the desired behaviour, because figure caption needs being translated and resides within the figure environment, while all other content of the environment typically consists of latex commands, which are hidden from translation anyway (thus making it redundant to hide the figure environment as a whole).
By this commit, figure environment is removed from the list of environments intended for being replaced by a token as a whole.
Note that the most essential command in the figure environment looking like "\includegraphics[width=0.5\textwidth]{fig1}" is processed correctly (totally hidden from translation) due to the previous commit which deals with latex commands having square-bracketed optional arguments.